RFID Based Smart Study Room
In this digital  Technology age, we find ourselves performing tasks manually in the study room, which is not only time-consuming but also inconvenient. For instance, we must manually control the lights and fan using Bluetooth technology. However, our projects using RFID technology, offering automatic opening of the study room door and displaying when a person enters the room. Instantly, the lights and fan turn on, providing a seamless experience for anyone entering the study. Furthermore, with the integration of Bluetooth, individuals can utilize their smart devices to control the lights and fan, enhancing convenience. Our innovative project aims to streamline processes, saving significant time and eliminating hassle. Gone are the days of manual operation; everything within the studyroom, including the fan and lights, will be automated. Additionally, we address the issue of wastage by incorporating a PIR sensor to automatically turn off the washroom lights when no one is present, thereby reducing our electricity bill.